tarot-backend/requirements.txt
fastapi>=0.104.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
uvicorn>=0.24.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.31.0
zhipuai>=2.0.0
openai>=1.0.0
PyJWT>=2.8.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
jinja2>=3.1.0
bcrypt>=4.0.0
email-validator>=2.0.0

